+++ packages/react/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## UNRELEASED
 
+## 4.0.12
+
 ## 4.0.11
 
 ### Changed